Fix issue with not writing final block marker
For certain values of BUFSIZE and BLOCKSIZE it was possible for the main
while loop to exit where size == 0 and we never wrote the final block
marker. Is really just coming from me complicating the logic by wanting
to support separate BUFSIZE and BLOCKSIZE. Also happens because feof()
doesn't return true until there was a read that returned 0. So if
BLOCKSIZE % BUFSIZE == 0, then would hit this case.